+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     as type
 
 * Added support for the field country as specified in APP-6 D
+* Added support for the field installationComposition as specified in APP-6 D
 
 ### Changed